build: move e2e tests out of routes to root folder
Ref. metriport/metriport-internal#1634
