+++ packages/api/src/routes/medical/__tests__/fhir-r4/e2e/organization.test.e2e.ts
@@ -1,5 +1,5 @@
 import { AxiosResponse } from "axios";
-import { api } from "../../../../__tests__/shared";
+import { api } from "../../../../../__tests__/e2e/shared";
 import { makeOrganization } from "./organization";
 
 jest.setTimeout(15000);
