+++ packages/api/src/routes/medical/__tests__/fhir-r4/e2e/organization.ts
@@ -1,4 +1,4 @@
-import { nanoid } from "../../../../__tests__/shared";
+import { nanoid } from "../../../../../__tests__/e2e/shared";
 
 const defaultId = "2.16.840.1.113883.3.9621.5." + nanoid();
